Move dbms operations into dbms sub-package
