+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/spi/TransactionalContextWrapper.scala
@@ -23,9 +23,10 @@ import org.neo4j.cypher.internal.compiler.v3_0.spi.QueryTransactionalContext
 import org.neo4j.graphdb.{Lock, PropertyContainer}
 import org.neo4j.kernel.GraphDatabaseQueryService
 import org.neo4j.kernel.api.KernelTransaction.Revertable
+import org.neo4j.kernel.api.dbms.DbmsOperations
 import org.neo4j.kernel.api.security.AccessMode
 import org.neo4j.kernel.api.txstate.TxStateHolder
-import org.neo4j.kernel.api.{DbmsOperations, ReadOperations, Statement}
+import org.neo4j.kernel.api.{ReadOperations, Statement}
 import org.neo4j.kernel.impl.query.TransactionalContext
 
 case class TransactionalContextWrapper(tc: TransactionalContext) extends QueryTransactionalContext {
