+++ community/kernel/src/main/java/org/neo4j/kernel/api/dbms/DbmsOperations.java
@@ -17,9 +17,10 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.api;
+package org.neo4j.kernel.api.dbms;
 
 import org.neo4j.collection.RawIterator;
+import org.neo4j.kernel.api.KernelAPI;
 import org.neo4j.kernel.api.exceptions.ProcedureException;
 import org.neo4j.kernel.api.proc.ProcedureSignature;
 import org.neo4j.kernel.api.security.AccessMode;
